medsync-backend/api/tests/test_jwt_algorithm.py
"""
JWT Algorithm Verification Tests

Ensures HS256 is used for single-backend authentication (safe).
Documents that if cross-hospital X-Consent-Token is added, RS256 MUST be used (security requirement).
"""

from django.test import TestCase
from django.conf import settings
from rest_framework_simplejwt.settings import api_settings as jwt_settings


class TestJWTAlgorithm(TestCase):
    """Verify JWT algorithm configuration for security."""

    def test_jwt_algorithm_explicitly_configured(self):
        """Verify that ALGORITHM is explicitly set (not defaulted)."""
        # Check that SIMPLE_JWT config includes ALGORITHM key
        self.assertIn("ALGORITHM", settings.SIMPLE_JWT, 
                     "ALGORITHM must be explicitly configured in SIMPLE_JWT settings")

    def test_jwt_algorithm_is_hs256_for_single_backend(self):
        """
        Verify that regular JWT tokens use HS256 (HMAC-SHA256).
        
        ✅ SAFE: HS256 is appropriate for single-backend scenarios where:
           - Only the backend has the shared secret
           - Backend both signs and verifies tokens
           - No cross-party verification needed
        
        ⚠️  UNSAFE for: Multi-party scenarios (e.g., Hospital A sends to Hospital B)
            Would require RS256 (asymmetric) instead
        """
        algorithm = jwt_settings.ALGORITHM
        self.assertEqual(
            algorithm, "HS256",
            f"Regular JWT tokens must use HS256 for single-backend verification. Got: {algorithm}"
        )

    def test_algorithm_configuration_documented(self):
        """
        Verify that SIMPLE_JWT includes a comment explaining the algorithm choice.
        
        This is a "soft" test - checks that the config exists and can be inspected in source.
        """
        # This test documents the security requirement
        # Actual comment is in medsync_backend/settings.py line 342
        config_str = str(settings.SIMPLE_JWT)
        self.assertIn("HS256", config_str)


class TestCrossHospitalSecurityRequirement(TestCase):
    """
    Document the RS256 requirement for cross-hospital scenarios.
    
    These tests are currently skipped because X-Consent-Token is NOT implemented.
    If X-Consent-Token is added in the future, these tests MUST pass.
    """

    def test_x_consent_token_must_use_rs256_not_hs256(self):
        """
        FUTURE TEST: When X-Consent-Token is implemented, it MUST use RS256.
        
        🔴 SECURITY REQUIREMENT:
        - HS256 (symmetric): Both parties have same secret → both can forge
        - RS256 (asymmetric): Only signer has private key → verifier cannot forge
        
        Example:
          Hospital A signs consent token with RS256 (private key)
          Hospital B receives token and verifies with RS256 (public key)
          ✅ Hospital B can verify authenticity
          ❌ Hospital B CANNOT forge new tokens (no private key)
          
        If Hospital B had HS256 shared secret:
          ✅ Hospital B can verify
          ❌ Hospital B CAN forge (knows the secret!)
          ❌ SECURITY BREACH
        """
        # This test documents the requirement and would be implemented when X-Consent-Token is added
        # When implemented, should verify:
        # 1. X-Consent-Token uses RS256 (not HS256)
        # 2. Private key is controlled by central platform only
        # 3. Public keys are distributed to hospitals for verification only
        # 4. Token cannot be forged by receiving hospital
        pass

    def test_current_cross_facility_uses_database_not_jwt(self):
        """
        Verify that current cross-facility access uses database queries (safer than JWT).
        
        ✅ CURRENT IMPLEMENTATION:
        - Access checks via database: Consent, Referral, BreakGlassLog models
        - Backend enforces all access control
        - No JWT tokens transmitted between hospitals
        - Cannot be replayed (checked on each access)
        - Immediately revocable
        
        This is actually SAFER than JWT tokens would be (if they used HS256).
        """
        from api.utils import can_access_cross_facility
        from interop.models import Consent
        
        # This test documents that current implementation is database-backed (safe)
        # The function can_access_cross_facility() queries Consent, Referral, BreakGlassLog
        # It does NOT use JWT tokens
        self.assertTrue(hasattr(Consent, "objects"))


class TestAlgorithmSecurityModel(TestCase):
    """
    Document the security model and why different algorithms are used/required.
    """

    def test_hs256_security_model_explained(self):
        """
        HS256 (HMAC-SHA256) Security Model:
        
        When Used (SAFE):
        - Single-backend systems where only backend has the secret
        - Example: Django app issues JWT, Django app verifies JWT
        - Backend is sole authority
        
        When Used (UNSAFE):
        - Multi-party systems where multiple parties have the secret
        - Example: Hospital A issues JWT, Hospital B verifies with same secret
        - Hospital B can forge tokens (knows the secret)
        
        Current MedSync Usage (SAFE):
        - Regular access tokens: HS256 (backend only, safe)
        - Cross-facility access: Database queries (backend enforces, safe)
        - No multi-party JWT tokens used
        """
        # This test documents the security model
        self.assertEqual(jwt_settings.ALGORITHM, "HS256")

    def test_rs256_security_model_requirement(self):
        """
        RS256 (RSA-SHA256) Security Model:
        
        Required For:
        - Multi-party verification where only signer should sign
        - Example: Central platform issues consent token, hospitals verify
        
        How It Works:
        - Signer (central platform): Has PRIVATE key
        - Verifiers (hospitals): Have PUBLIC key
        - Verifier can check signature but cannot forge (no private key)
        
        When X-Consent-Token is Added:
        - MUST use RS256 (not HS256)
        - Central platform keeps private key secure
        - Distribute public key to hospitals
        - Hospitals verify tokens but cannot forge
        
        If This Rule Is Violated:
        - Using HS256 for cross-hospital tokens
        - Hospitals receive shared secret
        - Hospitals can forge consent tokens
        - SECURITY BREACH
        """
        # This test documents the requirement
        # Should be enforced when X-Consent-Token is added
        pass


