medsync-backend/requirements-local.txt
# Local dev + CI: merged prod + dev (flat file; no -r includes — avoids broken third-party parsers, e.g. Vercel).
# Sync with repo root: ../requirements.txt (prod) + ../requirements-dev.txt (dev).

Django>=4.2,<5
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-otp>=1.4
pyotp>=2.9
python-decouple>=3.8
# WebAuthn/Passkey support
webauthn>=1.14
# Postgres (Neon / production): driver + DATABASE_URL parsing
psycopg[binary]>=3.1
dj-database-url>=2.1
redis>=5.0
Pillow>=10.0
reportlab>=4.0
channels>=4.0
channels-redis>=4.2
daphne>=4.0
scikit-learn>=1.3.0
xgboost>=2.0.0,<3
numpy>=1.24.0
pandas>=2.0.0
joblib>=1.3.0
faiss-cpu>=1.7.4
celery>=5.3.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0
kombu>=5.3.0
django-cryptography>=1.1

django-debug-toolbar>=4.2
pytest>=7.4
pytest-django>=4.5
crewai>=0.1.0
